refactor: Move mod compatibility code out of `main.lua`
